src/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
redis==3.5.3
faiss-cpu==1.7.1
sentence-transformers==2.1.0
torch==1.9.0
transformers==4.11.3
nltk==3.6.3
pydantic==1.8.2